pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pca-estafette"
version = "0.1.0"

requires-python = ">= 3.13"
dependencies = [
    "pydantic>=2.11",
    "python-dotenv>=1.1",
    "rich>=14.0",
    "sib-api-v3-sdk>=7.6",
    "typer>=0.15",
    "boto3>=1.34",
]

license = {file = "LICENSE"}
readme = "README.md"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]

authors = [
  {name = "Etienne Madinier"},
  {name = "Selim Mellouk-Halgand"}
]
maintainers = [
  {name = "Etienne Madinier"},
  {name = "Selim Mellouk-Halgand"}
]

[project.urls]
Homepage = "https://pointcarre.app"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.package-data]
scenery = ["LICENSE"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["estafettes*"]  # Import as 'estafettes'

[tool.pydocstyle]
convention = "google"
match = ".*\\.py"